[62908364] Added NGINX
The gds nginx module was included in librarian puppet and then configured via the common.yaml hiera file. There is an instance of the nginx::vhost::redirect definition in the nodes.pp file to set up the redirect for gov.uk to https://www.gov.uk. README was also updated to specify what needs doing.