[62908364] - Adding Hiera-eyaml-gpg Config (Not quite working)
Adding hiera-eyaml-gpg config. Following installation of the included gems in the gemfile, config is requred in both the hiera.yaml and common.yaml file. Before that, gpg keys must be created without a passphrase and stored in a secure location for decryption. Using this gpg, the ssl data can be encrypted and placed into the yaml files which *should* be decrypted on the fly. Nginx is also included with the relevant redirect logic. There are two current bugs: 1) The encrypted hiera data isn't being decrypted on puppet runs. 2) The path for ssl keys and certs isn't being pulled in correctly.