[62908364] Productionise build
Productionised the build by specifying nginx version, adding gds accounts, adding the ppa, adding sudo, clamav, ufw, apt etc. The repo was also modified to the newer shell provisioned vagrant setup and an environment was created to configure sudo and accounts for vagrant. The list of gpg recipients was also added to hieradata so that the encryption process is smoother.